Use class (no struct) for vars and write TTree
To ease the creation of the output TTree a class has been
created instead of using a struct and manually asigning
branches. In order to have ROOT I/O, dictionary creation
for the class (EventData) has also been automated. The
name of the file is set with a TSelector option "ofile="
which is parsed from -o or -outout when using process_tchain
executable.