Comment all TTree vars not used (automatically)
Created a basic script (grep, cut and sed) that comments lines that
contain a variable of the TTree that is not used in
the TSelector implementation. This will greatly reduce
the amount of data to read ( from 15Kb to 3 Kb) speeding
the skimming process. In the furure this will not be
required if ROOT6 TTreeReaderValue<T> and TTreeReaderArray<T>
are used.